refactor restore cache state again
